pyproject.toml
[tool.poetry]
name = "flask-psql-service"
version = "0.1.0"
description = ""
authors = ["TravisAucoin <travis_aucoin@stuller.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.3"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
psycopg2 = "^2.9.3"
alembic = "1.4.0"
python-dotenv = "^0.20.0"
Flask-Login = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

